+++ tests/autotuner/test_trtllm_fused_moe_autotuner_integration.py
@@ -561,104 +561,6 @@ def test_fp8_block_scale_moe_fused_shared_experts_autotune(num_fused_shared_expe
         )
 
 
-@pytest.mark.parametrize("num_fused_shared_experts", [1, 2])
-def test_fp4_block_scale_moe_fused_shared_experts_autotune(num_fused_shared_experts):
-    """FP4 fused shared experts must agree with the autotuner on the fused dimensions.
-
-    FP4 sibling of test_fp8_block_scale_moe_fused_shared_experts_autotune:
-    ``get_valid_tactics`` must query ``trtllm_get_valid_moe_configs`` with the
-    fused (routed + shared) ``top_k`` / ``local_num_experts`` so the chosen
-    tactic passes ``prepare_moe()``'s ``effectiveTopK`` / ``effectiveLocalExperts``
-    validation. Weight contents are uninitialized (numerics are irrelevant) —
-    smoke test: must not crash during autotuning.
-    """
-    _require_sm100()
-    reset_autotuner()
-    device = torch.device("cuda:0")
-
-    from flashinfer.fused_moe import trtllm_fp4_block_scale_moe
-
-    num_experts = 256  # routed experts
-    top_k = 8
-    n_group = 8
-    topk_group = 4
-    hidden_size = 1024
-    intermediate_size = 512
-    num_tokens = 8
-    # Weight tensors carry the routed + fused shared experts in the expert dim.
-    total_experts = num_experts + num_fused_shared_experts
-
-    routing_logits = torch.rand(
-        num_tokens, num_experts, dtype=torch.float32, device=device
-    )
-    routing_bias = torch.randn(num_experts, dtype=torch.bfloat16, device=device)
-    hidden_states = torch.randn(
-        num_tokens, hidden_size, dtype=torch.bfloat16, device=device
-    )
-    gemm1_weights = torch.empty(
-        total_experts,
-        intermediate_size * 2,
-        hidden_size // 2,
-        dtype=torch.uint8,
-        device=device,
-    )
-    gemm1_weights_scale = torch.ones(
-        total_experts,
-        intermediate_size * 2,
-        hidden_size // 2 // 16,
-        dtype=torch.float32,
-        device=device,
-    ).to(torch.float8_e4m3fn)
-    gemm2_weights = torch.empty(
-        total_experts,
-        hidden_size,
-        intermediate_size // 2,
-        dtype=torch.uint8,
-        device=device,
-    )
-    gemm2_weights_scale = torch.ones(
-        total_experts,
-        hidden_size,
-        intermediate_size // 2 // 16,
-        dtype=torch.float32,
-        device=device,
-    ).to(torch.float8_e4m3fn)
-    output = torch.empty(num_tokens, hidden_size, dtype=torch.bfloat16, device=device)
-
-    with autotune(tune_mode=True):
-        trtllm_fp4_block_scale_moe(
-            routing_logits=routing_logits,
-            routing_bias=routing_bias,
-            hidden_states=hidden_states,
-            hidden_states_scale=None,
-            gemm1_weights=gemm1_weights,
-            gemm1_weights_scale=gemm1_weights_scale,
-            gemm1_bias=None,
-            gemm1_alpha=None,
-            gemm1_beta=None,
-            gemm1_clamp_limit=None,
-            gemm2_weights=gemm2_weights,
-            gemm2_weights_scale=gemm2_weights_scale,
-            gemm2_bias=None,
-            output1_scale_scalar=None,
-            output1_scale_gate_scalar=None,
-            output2_scale_scalar=None,
-            num_experts=num_experts,
-            top_k=top_k,
-            n_group=n_group,
-            topk_group=topk_group,
-            intermediate_size=intermediate_size,
-            local_expert_offset=0,
-            local_num_experts=num_experts,
-            routed_scaling_factor=2.5,
-            routing_method_type=RoutingMethodType.DeepSeekV3.value,
-            do_finalize=True,
-            output=output,
-            tune_max_num_tokens=8,
-            num_fused_shared_experts=num_fused_shared_experts,
-        )
-
-
 @pytest.mark.parametrize(
     "invalid_tactic",
     [
